API Reference

PkgFactory.CLI

PkgFactory.check_file_on_github

PkgFactory.check_repo

PkgFactory.check_status_code

PkgFactory.commit_file_on_github

PkgFactory.commit_files_on_github

PkgFactory.create_branch_gh_pages

PkgFactory.create_repo

PkgFactory.device_flow_begin

PkgFactory.device_flow_end

PkgFactory.generate_template_dict

PkgFactory.get_codecov_url

PkgFactory.get_file_on_github

PkgFactory.get_user_domain

PkgFactory.list_files

PkgFactory.load_file

PkgFactory.set_codecov

PkgFactory.set_deploy_key

PkgFactory.set_repository_secret

PkgFactory.update_template

PkgFactory.verify_package_name

PkgFactory.CLI — Method

CLI()

PkgFactory.check_file_on_github — Method

check_file_on_github(
    access_token::String,
    owner_name::String,
    repo_name::String,
    branch_name::String,
    repo_path::String
) -> Bool

PkgFactory.check_repo — Method

check_repo(
    access_token::String,
    owner_name::String,
    repo_name::String
) -> Bool

PkgFactory.check_status_code — Method

check_status_code() -> Bool

PkgFactory.commit_file_on_github — Method

commit_file_on_github(
    access_token::String,
    owner_name::String,
    repo_name::String,
    branch_name::String,
    commit_message::String,
    path::String,
    content::String
)

PkgFactory.commit_files_on_github — Method

Warning

This function does not work for initial commit.

Signature:

commit_files_on_github(
    access_token::String,
    owner_name::String,
    repo_name::String,
    branch_name::String,
    commit_message::String,
    paths_and_contents::Dict{String, String}
) -> GitHub.Reference

Example:

access_token = "YOUR_ACCESS_TOKEN"
owner_name = "user"
repo_name = "my-github-repo"
branch_name = "main"
commit_message = "Update hello1.md and hello2.md"
paths_and_contents = Dict(
    "hello1.md" => "Hello, 1",
    "hello2.md" => "Hello, 2",
)

PkgFactory.commit_files_on_github(
    access_token,
    owner_name,
    repo_name,
    branch_name,
    commit_message,
    paths_and_contents,
)

PkgFactory.create_branch_gh_pages — Method

create_branch_gh_pages(
    access_token::String,
    owner_name::String,
    repo_name::String
) -> Bool

PkgFactory.create_repo — Method

create_repo(
    access_token::String,
    owner_name::String,
    repo_name::String,
    package_description::String
) -> Union{Nothing, Bool, Float64, Int64, String, JSON3.Array, JSON3.Object}

PkgFactory.device_flow_begin — Method

device_flow_begin(client_id::String) -> JSON3.Object

PkgFactory.device_flow_end — Method

device_flow_end(
    client_id::String,
    device_code::String
) -> Union{Nothing, Bool, Float64, Int64, String, JSON3.Array, JSON3.Object}

PkgFactory.generate_template_dict — Method

generate_template_dict(
    owner_name::String,
    repo_name::String,
    author_names::Vector{String},
    package_description::String
) -> Dict{String, String}

PkgFactory.get_codecov_url — Method

get_codecov_url(
    owner_name::String,
    repo_name::String
) -> String

PkgFactory.get_file_on_github — Method

get_file_on_github(
    access_token::String,
    owner_name::String,
    repo_name::String,
    repo_path::String
) -> GitHub.Content

PkgFactory.get_user_domain — Method

get_user_domain(access_token::String) -> Any

PkgFactory.list_files — Method

list_files() -> Vector{String}

PkgFactory.load_file — Method

load_file(path::String) -> String

PkgFactory.set_codecov — Method

set_codecov(
    access_token::String,
    owner_name::String,
    repo_name::String,
    codecov_token::String
) -> Bool

PkgFactory.set_deploy_key — Method

set_deploy_key(
    access_token::String,
    owner_name::String,
    repo_name::String
) -> Bool

PkgFactory.set_repository_secret — Method

set_repository_secret(
    access_token::String,
    owner_name::String,
    repo_name::String,
    secret_name::String,
    secret_value::String
) -> Bool

PkgFactory.update_template — Method

Warning

This function is only for developers. This function updates the template files in the PkgFactory.jl/template directory.

Signature:

update_template(
    owner_name::String,
    repo_name::String,
    author_names::Vector{String}
)

Example:

PkgFactory.update_template("OWNER_NAME", "template", ["AUTHOR1", "AUTHOR2"])

PkgFactory.verify_package_name — Method

This function verifies the package name according to the package naming rules of Pkg.jl and the automatic merging guidelines of RegistryCI.jl.

verify_package_name(package_name::String) -> String